P4-2228 change move and journey to have Location type rather than string
This is so that we can easily get the sitename and location type.

Achieved using LocationParser.kt which looks at a list of locations.
If the location name is null it returns null
If the location for the nomis agency id exists, it uses this
If the location doesn't exist, it creates a new one with the agency id as sitename and locationtype unknown.

Remove location mapping from PriceCalculator.kt as it's no longer needed.